feat: server-side admin auth check
