+++ WT4Q/src/app/admin-login/AdminLoginClient.tsx
@@ -1,0 +1,145 @@
+'use client';
+
+import { FC, useState, FormEvent, useTransition } from 'react';
+import { useRouter } from 'next/navigation';
+import styles from './AdminLogin.module.css';
+import { API_ROUTES } from '@/lib/api';
+
+interface LoginRequest {
+  email: string;
+  password: string;
+}
+
+const EMAIL_REGEX = /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/;
+
+const AdminLoginClient: FC = () => {
+  const router = useRouter();
+  const [email, setEmail] = useState<string>('');
+  const [password, setPassword] = useState<string>('');
+  const [error, setError] = useState<string | null>(null);
+  const [emailError, setEmailError] = useState<string | null>(null);
+  const [showPassword, setShowPassword] = useState<boolean>(false);
+  const [isPending, startTransition] = useTransition();
+  const isLoading = isPending;
+
+
+  const validate = (): boolean => {
+    let valid = true;
+    if (!EMAIL_REGEX.test(email)) {
+      setEmailError('Please enter a valid email address');
+      valid = false;
+    } else {
+      setEmailError(null);
+    }
+    return valid;
+  };
+
+  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    setError(null);
+
+    if (!validate()) return;
+
+    startTransition(async () => {
+      try {
+        const response = await fetch(API_ROUTES.ADMIN_AUTH.LOGIN, {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/json' },
+          credentials: 'include',
+          body: JSON.stringify({ email, password } as LoginRequest),
+        });
+        const data: { token?: string; message?: string } = await response.json();
+
+        if (!response.ok ) {
+          throw new Error(data.message || 'Login failed');
+        }
+
+        router.replace('/admin/dashboard');
+      } catch (err) {
+        if (err instanceof Error) {
+          setError(err.message);
+        } else {
+          setError('An unexpected error occurred');
+        }
+      }
+    });
+  };
+
+  return (
+    <main className={styles.container}>
+      <section className={styles.card} aria-labelledby="admin-login-title">
+        <h1 id="admin-login-title" className={styles.title}>
+          Admin Login
+        </h1>
+        {error && (
+          <div role="alert" className={styles.error}>
+            {error}
+          </div>
+        )}
+        <form onSubmit={handleSubmit} className={styles.form} noValidate>
+          <div className={styles.field}>
+            <label htmlFor="email" className={styles.label}>
+              Email
+            </label>
+            <input
+              id="email"
+              type="email"
+              autoComplete="username"
+              autoFocus
+              required
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              className={styles.input}
+              aria-invalid={!!emailError}
+              aria-describedby={emailError ? 'email-error' : undefined}
+            />
+            {emailError && (
+              <p id="email-error" className={styles.fieldError} role="alert">
+                {emailError}
+              </p>
+            )}
+          </div>
+
+          <div className={styles.field}>
+            <label htmlFor="password" className={styles.label}>
+              Password
+            </label>
+            <div className={styles.passwordWrapper}>
+              <input
+                id="password"
+                type={showPassword ? 'text' : 'password'}
+                autoComplete="current-password"
+                required
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+                className={styles.input}
+              />
+              <button
+                type="button"
+                onClick={() => setShowPassword((prev) => !prev)}
+                className={styles.toggleButton}
+                aria-label={showPassword ? 'Hide password' : 'Show password'}
+              >
+                {showPassword ? '🙈' : '👁️'}
+              </button>
+            </div>
+          </div>
+
+          <button
+            type="submit"
+            disabled={isLoading}
+            className={styles.button}
+          >
+            {isLoading ? (
+              <span className={styles.spinner} aria-hidden="true" />
+            ) : (
+              'Sign In'
+            )}
+          </button>
+        </form>
+      </section>
+    </main>
+  );
+};
+
+export default AdminLoginClient;
